[fix] Revamp `preferredLayouts` state for `set_page_config` and layout in User Settings (#9723)
## Describe your changes

Update:
In this update, we revamped the `preferredLayouts` state in `App.tsx`
(previously `pageLayouts`) and fixed a number of issues with
1. apps flickering (when not calling `set_page_config`) 
2. apps having fragments rerunning
3. wide mode in User Settings

Here's a breakdown of the implementation by functions/files:
1. `handlePageConfigChanged` & `saveSetting`: `set_page_config` and User
Settings panel are the 2 only places that could change the layout. As a
result, we save the layout value to `preferredLayouts` in these 2
handlers
2. `handleNewSession` (not changed): For every rerun, we apply
previously saved `preferredLayouts` of the corresponding page. If none
matches, apply `CENTERED`
3. `handleNavigation`: If we get a new page script hash from navigation
which has never been initialized in `preferredLayouts`, we carry over
the layout from previous page for the new one.

--

~~We are seeing an unexpected behavior such that when an MPA page
containing fragments reruns, the page layout would change to centered.
This is an [issue reported in our
forum](https://discuss.streamlit.io/t/bad-setin-index-issue-with-using-fragments-that-do-not-render-anything-within-multipage-app/82675/9?u=raethlein)
flagged by @raethlein. Upon investigation, we have identified that the
fix in #9479 introduced this behavior change.~~

~~The root cause is that after a triggered rerun the page script hash
changes, causing the `pageLayouts` object state not recognizing the new
hash value and assigning the default/centered layout.
Our proposed fix here is to only perform layout update when there is no
fragment running.~~

~~I also added new e2e test cases with fragments to verify the fix.~~

## GitHub Issue Link (if applicable)
[Post in discuss
forum](https://discuss.streamlit.io/t/bad-setin-index-issue-with-using-fragments-that-do-not-render-anything-within-multipage-app/82675?u=raethlein)

## Testing Plan

- E2E Tests: new test case in MPAv2 that mocks fragment rerunning

### Manually testing app reported in forum
#### Before


https://github.com/user-attachments/assets/41c3f629-5f69-42c3-9e5f-e0fec9d1bff3

#### After


https://github.com/user-attachments/assets/77de515e-2e89-47d4-b49d-c03126d61f41


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.